LPS-146600 Task is a kind of Node, create a BaseDefaultWorkflowNode to avoid duplications
