+++ portal-kernel/src/com/liferay/portal/kernel/workflow/BaseDefaultWorkflowNode.java
@@ -1,0 +1,80 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.kernel.workflow;
+
+import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.util.HtmlUtil;
+import com.liferay.portal.kernel.util.LocaleUtil;
+
+import java.util.Locale;
+import java.util.Map;
+
+/**
+ * @author Feliphe Marinho
+ */
+public abstract class BaseDefaultWorkflowNode implements WorkflowNode {
+
+	@Override
+	public String getLabel(Locale locale) {
+		String label = _labelMap.get(locale);
+
+		if (label != null) {
+			return HtmlUtil.escape(label);
+		}
+
+		label = _labelMap.get(LocaleUtil.getSiteDefault());
+
+		if (label != null) {
+			return HtmlUtil.escape(label);
+		}
+
+		Language language = LanguageUtil.getLanguage();
+
+		return HtmlUtil.escape(language.get(locale, _name));
+	}
+
+	@Override
+	public Map<Locale, String> getLabelMap() {
+		return _labelMap;
+	}
+
+	@Override
+	public String getName() {
+		return _name;
+	}
+
+	@Override
+	public Type getType() {
+		return _type;
+	}
+
+	public void setLabelMap(Map<Locale, String> labelMap) {
+		_labelMap = labelMap;
+	}
+
+	public void setName(String name) {
+		_name = name;
+	}
+
+	public void setType(Type type) {
+		_type = type;
+	}
+
+	private Map<Locale, String> _labelMap;
+	private String _name;
+	private Type _type;
+
+}
