+++ portal-kernel/src/com/liferay/portal/kernel/workflow/DefaultWorkflowNode.java
@@ -14,55 +14,8 @@
 
 package com.liferay.portal.kernel.workflow;
 
-import com.liferay.portal.kernel.language.LanguageUtil;
-
-import java.util.Locale;
-import java.util.Map;
-
 /**
  * @author Feliphe Marinho
  */
-public class DefaultWorkflowNode implements WorkflowNode {
-
-	public String getLabel(Locale locale) {
-		String label = _labelMap.get(locale);
-
-		if (label != null) {
-			return label;
-		}
-
-		return LanguageUtil.get(locale, _name);
-	}
-
-	@Override
-	public Map<Locale, String> getLabelMap() {
-		return _labelMap;
-	}
-
-	@Override
-	public String getName() {
-		return _name;
-	}
-
-	@Override
-	public Type getType() {
-		return _type;
-	}
-
-	public void setLabelMap(Map<Locale, String> labelMap) {
-		_labelMap = labelMap;
-	}
-
-	public void setName(String name) {
-		_name = name;
-	}
-
-	public void setType(Type type) {
-		_type = type;
-	}
-
-	private Map<Locale, String> _labelMap;
-	private String _name;
-	private Type _type;
-
+public class DefaultWorkflowNode extends BaseDefaultWorkflowNode {
 }
