+++ portal-kernel/src/com/liferay/portal/kernel/workflow/WorkflowNode.java
@@ -22,6 +22,8 @@ import java.util.Map;
  */
 public interface WorkflowNode {
 
+	public String getLabel(Locale locale);
+
 	public Map<Locale, String> getLabelMap();
 
 	public String getName();
